Rename files
Again, not really serialization
